Refactor event handlers and modularize UI logic
Extracted and modularized event handlers into dedicated files, improving maintainability and readability. Specific actions, such as modal operations, character management, and UI event handling, are now separated into their respective modules.
