+++ src/events/global-events.js
@@ -1,5 +1,9 @@
 import { initializeCharDetailsEvents } from "./characterDetails-events.js";
 import { initializeModalEvents } from "./modal-events.js";
+import { initializeUIMenuEvents } from "./UImenu-events.js";
+import { initializeCharactersEvents } from "./characters-events.js";
+import { initializeToolbarEvents } from "./toolbar-events.js";
+import { initializePresetsEvents } from "./presets-events.js";
 
 
 /**
@@ -8,5 +12,9 @@ import { initializeModalEvents } from "./modal-events.js";
 export function initializeEventHandlers() {
     // Initialize event handlers for different components
     initializeCharDetailsEvents();
+    initializeCharactersEvents();
     initializeModalEvents();
+    initializeUIMenuEvents();
+    initializeToolbarEvents();
+    initializePresetsEvents();
 }
